use helpers to make token mint
