+++ labs/create-token-mint-with-metadata.ts
@@ -1,29 +1,11 @@
-// Adapted from https://github.com/solana-labs/solana-program-library/blob/master/token/js/examples/metadata.ts
 import "dotenv/config";
 import {
-  clusterApiUrl,
-  Connection,
-  Keypair,
-  sendAndConfirmTransaction,
-  SystemProgram,
-  Transaction,
-} from "@solana/web3.js";
-import {
-  createInitializeMetadataPointerInstruction,
-  createInitializeMintInstruction,
-  ExtensionType,
-  getMintLen,
-  LENGTH_SIZE,
-  TOKEN_2022_PROGRAM_ID,
-  TYPE_SIZE,
-} from "@solana/spl-token";
-import type { TokenMetadata } from "@solana/spl-token-metadata";
-import { createInitializeInstruction, pack } from "@solana/spl-token-metadata";
-import {
+  airdropIfRequired,
   getExplorerLink,
   getKeypairFromEnvironment,
+  makeTokenMint,
 } from "@solana-developers/helpers";
-
+import { Connection, LAMPORTS_PER_SOL, clusterApiUrl } from "@solana/web3.js";
 const connection = new Connection(clusterApiUrl("devnet"));
 
 const user = getKeypairFromEnvironment("SECRET_KEY");
@@ -32,65 +14,51 @@ console.log(
   `🔑 Loaded our keypair securely, using an env file! Our public key is: ${user.publicKey.toBase58()}`
 );
 
-const mint = Keypair.generate();
-const DECIMALS = 9;
+const mintAuthority = user;
 
-const metadata: TokenMetadata = {
-  mint: mint.publicKey,
-  name: "Solana training token",
-  symbol: "TRAINING",
-  uri: "https://raw.githubusercontent.com/solana-developers/professional-education/main/labs/sample-token-metadata.json",
-  additionalMetadata: [["new-field", "new-value"]],
+await airdropIfRequired(
+  connection,
+  mintAuthority.publicKey,
+  100 * LAMPORTS_PER_SOL,
+  1 * LAMPORTS_PER_SOL
+);
+
+// TODO: what name and symbol are used in explorer, the one from the mint or the one from the metadata?
+// Remove whatever is not used
+// ANSWER: name from json metadata is used in explorer
+// what happens if we remove it?
+// TODO: push to github and see what happens
+const name = "Developer Education";
+const symbol = "EDU";
+const decimals = 9;
+const uri =
+  "https://raw.githubusercontent.com/solana-developers/professional-education/main/labs/sample-token-metadata.json";
+
+const additionalMetadata = {
+  shlerm: "frobular",
+  glerp: "flerpy",
+  gurperderp: "erpy",
+  nurmagerd: "flerpy",
+  zurp: "flerpy",
+  eruper: "flerpy",
+  zerperurperserp: "flerpy",
+  zherp: "flerpy",
 };
 
-// Work out how much SOL we need to store our Token
-const mintLength = getMintLen([ExtensionType.MetadataPointer]);
-const metadataLength = TYPE_SIZE + LENGTH_SIZE + pack(metadata).length;
-const mintLamports = await connection.getMinimumBalanceForRentExemption(
-  mintLength + metadataLength
+const mintAddress = await makeTokenMint(
+  connection,
+  mintAuthority,
+  name,
+  symbol,
+  decimals,
+  uri,
+  additionalMetadata
 );
 
-// make instructtons to:
-// - create the mint account,
-// - write the dnecessary data for metadata pointer,
-// - initialize the
-const mintTransaction = new Transaction().add(
-  SystemProgram.createAccount({
-    fromPubkey: user.publicKey,
-    newAccountPubkey: mint.publicKey,
-    space: mintLength,
-    lamports: mintLamports,
-    programId: TOKEN_2022_PROGRAM_ID,
-  }),
-  createInitializeMetadataPointerInstruction(
-    mint.publicKey,
-    user.publicKey,
-    mint.publicKey,
-    TOKEN_2022_PROGRAM_ID
-  ),
-  createInitializeMintInstruction(
-    mint.publicKey,
-    DECIMALS,
-    user.publicKey,
-    null,
-    TOKEN_2022_PROGRAM_ID
-  ),
-  createInitializeInstruction({
-    programId: TOKEN_2022_PROGRAM_ID,
-    mint: mint.publicKey,
-    metadata: mint.publicKey,
-    name: metadata.name,
-    symbol: metadata.symbol,
-    uri: metadata.uri,
-    mintAuthority: user.publicKey,
-    updateAuthority: user.publicKey,
-  })
+const explorerLink = getExplorerLink(
+  "address",
+  mintAddress.toBase58(),
+  "devnet"
 );
-const signature = await sendAndConfirmTransaction(connection, mintTransaction, [
-  user,
-  mint,
-]);
 
-const link = getExplorerLink("transaction", signature, "devnet");
-
-console.log(`✅ Success! Created token mint: ${link}`);
+console.log(`🏦 Token mint created at: ${explorerLink}`);
